+++ docs/benchmarks/benchmark_figure.py
@@ -100,10 +100,8 @@ for prob_name in ["narrow-insurance", "intermediate-insurance", "intermediate-ho
         )
         plot_df = plot_df.pivot(columns="library_name")
         plot_df.columns = plot_df.columns.get_level_values(1)
+        plot_df = plot_df.sort_index(axis=1).rename(columns={"r-glmnet": "glmnet"})
         plot_df.index = [x.title() for x in plot_df.index]
-        plot_df = plot_df[["h2o", "glum", "r-glmnet"]].rename(
-            columns={"r-glmnet": "glmnet"}
-        )
 
         title = prob_name.title() + "-" + ("Lasso" if reg == "lasso" else "Ridge")
         plot_df.plot.bar(
@@ -170,6 +168,7 @@ for prob_name in ["wide-insurance"]:
         )
         plot_df = plot_df.pivot(columns="library_name")
         plot_df.columns = plot_df.columns.get_level_values(1)
+        plot_df = plot_df.sort_index(axis=1).rename(columns={"r-glmnet": "glmnet"})
         plot_df.index = [x.title() for x in plot_df.index]
 
         title = prob_name.title() + "-" + ("Lasso" if reg == "lasso" else "Ridge")
@@ -237,6 +236,7 @@ plot_df = (
 )
 plot_df = plot_df.pivot(columns="library_name")
 plot_df.columns = plot_df.columns.get_level_values(1)
+plot_df = plot_df.sort_index(axis=1).rename(columns={"r-glmnet": "glmnet"})
 plot_df.index = [x.title() for x in plot_df.index]
 
 plot_df.plot.bar(